The ninth commit
There create new branch, it is working version of app, then we try merge it.

There done scrollable fragment_the_home_page.xml check comments for know how it is working (activity_for_fragments).

After that we gonna add new selector for another icons in menu, try to do best realization for fragments, cause it is work not well in my opinion.

In the future we maybe try to do something seems like navigation, but i not know how it is do.